+++ src/Pagination/ConfigurationBagConfigurator.php
@@ -49,7 +49,7 @@ final class ConfigurationBagConfigurator
 		$bag = new FilterBag();
 		
 		foreach ($filters as $key => $filter) {
-			$bag->set($key, new Filter($filter['property_name'], $filter['cast_type'], $filter['searchable'], $filter['ordering']));
+			$bag->set($key, new Filter($filter['property_name'], $filter['cast_type'], $filter['searchable'], $filter['orderable']));
 		}
 		
 		return $bag;